Switch argument declaration to rvalue ref
`Propagators::define_and_install_table` uses a lambda that captures an
`ExtensionalData` with members move-constructed with objects one of
which this function just received as a reference.

I believe there is a potential for a hidden undefined behaviour
(including SIGSEGV) if `vars` is used outside the function
after it is called, and the lambda gets deconstructed, along with the
underlying data moved into it. This does not currently happen but it
would be good to enforce this 'contract' with the compiler to protect
from future changes, by only accepting rvalue references for objects
which are to be moved into the lambda.

The alternative would be for `ExtensionalData` to only hold references,
with some thought about the guarantees for the underlying objects not
being deconstructed before the `ExtensionalData`.